DX-1288: Rename regex_capture to regex_capture_filter
