+++ lib/regex_capture_filter.rb
