+++ lib/swedbank-pay-design-guide-jekyll-theme.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
 require_relative 'sidebar'
-require_relative 'regex_capture'
+require_relative 'regex_capture_filter'
